Cleanup UTF7Encoding negative tests
- Make EncodingHelpers.cs more extensible
