+++ src/System.Text.Encoding/tests/UTF7Encoding/UTF7EncodingGetString.cs
@@ -37,62 +37,5 @@ namespace System.Text.Tests
             UTF7Encoding utf7 = new UTF7Encoding();
             string str = utf7.GetString(bytes, startIndex, count);
         }
-
-        // NegTest1: ArgumentNullException is not thrown when bytes is a null reference
-        [Fact]
-        public void NegTest1()
-        {
-            Byte[] bytes = null;
-            UTF7Encoding utf7 = new UTF7Encoding();
-            Assert.Throws<ArgumentNullException>(() =>
-            {
-                string str = utf7.GetString(bytes, 0, 2);
-            });
-        }
-
-        // NegTest2: ArgumentOutOfRangeException is not thrown when index is less than zero.
-        [Fact]
-        public void NegTest2()
-        {
-            Byte[] bytes = new Byte[] {
-                             85,  84,  70,  56,  32,  69, 110,
-                             99, 111, 100, 105, 110, 103,  32,
-                             69, 120,  97, 109, 112, 108, 101};
-            UTF7Encoding utf7 = new UTF7Encoding();
-            Assert.Throws<ArgumentOutOfRangeException>(() =>
-            {
-                string str = utf7.GetString(bytes, -1, bytes.Length);
-            });
-        }
-
-        // NegTest3: ArgumentOutOfRangeException is not thrown when count is less than zero
-        [Fact]
-        public void NegTest3()
-        {
-            Byte[] bytes = new Byte[] {
-                             85,  84,  70,  56,  32,  69, 110,
-                             99, 111, 100, 105, 110, 103,  32,
-                             69, 120,  97, 109, 112, 108, 101};
-            UTF7Encoding utf7 = new UTF7Encoding();
-            Assert.Throws<ArgumentOutOfRangeException>(() =>
-            {
-                string str = utf7.GetString(bytes, 0, -1);
-            });
-        }
-
-        // NegTest4: ArgumentOutOfRangeException is not thrown when index and count do not denote a valid range in bytes.
-        [Fact]
-        public void NegTest4()
-        {
-            Byte[] bytes = new Byte[] {
-                             85,  84,  70,  56,  32,  69, 110,
-                             99, 111, 100, 105, 110, 103,  32,
-                             69, 120,  97, 109, 112, 108, 101};
-            UTF7Encoding utf7 = new UTF7Encoding();
-            Assert.Throws<ArgumentOutOfRangeException>(() =>
-            {
-                string str = utf7.GetString(bytes, 1, bytes.Length);
-            });
-        }
     }
 }
